// src/navigations/domains/types/app.types.ts
// types

// app types and enum
type AppStackParamList = {
    Map: undefined;
    Workflow: undefined;
    Exit: undefined;
    Infographics: undefined;
    Eta: undefined;
    Road: undefined;
    Dominating: undefined;
    Speed: undefined;
    Onboard: undefined;
}

enum AppEnum {
    Map = "Map",
    Workflow = "Workflow",
    Exit = "Exit",
    Infographics = 'Infographics',
    Eta = "Eta",
    Road = "Road",
    Dominating = "Dominating",
    Speed = "Speed",
    Onboard = "Onboard"
}

export { AppStackParamList, AppEnum };